requirements.txt
langchain
langgraph
langchain_community
langchain_core
langchain_groq
fastapi
uvicorn
watchdog
python-dotenv